Relocate persistence/logger systems out of app/util/ (closes #1197)
Per #1197, app/util/ should hold only header-inline pure functions.
Three files were stateful systems-in-costume:

- persistence_policy.{h,cpp}  → systems/persistence_policy_system.{h,cpp}
- session_logger.{h,cpp}      → systems/session_logger_system.{h,cpp}
- high_score_persistence.{h,cpp} → systems/high_score_system.{h,cpp}

session_logger.cpp owns FILE* lifetime and reacts to EnTT signals;
persistence_policy is shared filesystem code with structured status;
high_score_persistence is the high-score system's behavior layer.

Updates: include paths in all callers (app + tests), the
SHAPESHIFTER_SOURCE_DIR string lookups in
test_settings_persistence and test_wasm_beforeunload_lifecycle, and
the relocated paths in design-docs/tester-personas-tech-spec.md.

app/util/ now contains only the reference-shape header-inline utils:
lane_utils.h, rhythm_math.h, shape_lane_mapping.h.

Build clean, all 2956 assertions / 901 test cases pass.
