+++ app/entities/settings.h
@@ -6,7 +6,7 @@
 #include <filesystem>
 #include <string>
 
-#include "../util/persistence_policy.h"
+#include "../systems/persistence_policy_system.h"
 #include "tags/tags.h"
 
 struct SettingsState {
