+++ app/systems/session_logger_system.cpp
@@ -1,4 +1,4 @@
-#include "session_logger.h"
+#include "session_logger_system.h"
 #include "../components/obstacle.h"
 #include "../components/rhythm.h"
 #include "../components/scoring.h"
