+++ tests/test_session_logger.cpp
@@ -1,6 +1,6 @@
 #include <catch2/catch_test_macros.hpp>
 
-#include "util/session_logger.h"
+#include "systems/session_logger_system.h"
 
 TEST_CASE("session log frame advances from central game-loop hook", "[session_logger][issue112]") {
     SessionLog log{};
