+++ tests/test_test_player_system.cpp
@@ -3,7 +3,7 @@
 #include "components/test_player.h"
 #include "content/level_content_config.h"
 #include "session/test_player_session.h"
-#include "util/session_logger.h"
+#include "systems/session_logger_system.h"
 #include <cstdio>
 #include <string>
 static entt::registry make_test_player_registry(TestPlayerSkill skill = TestPlayerSkill::Pro) {
